move beta js into modules
