+++ platform/src/js/campfire-theme.js
@@ -4,7 +4,7 @@ import $ from 'jquery'
 import showdown from 'showdown'
 import purify from 'dompurify'
 import finallycomments from 'finallycomments'
-import beta from '../beta.js'
+import beta from './modules/finally-try'
 beta.init()
 
 const converter = new showdown.Converter({ tables: true })
