+++ platform/src/js/lens-theme.js
@@ -5,7 +5,7 @@ import showdown from 'showdown'
 import inview from 'jquery-inview'
 import finallycomments from 'finallycomments'
 import purify from 'dompurify'
-import beta from '../beta.js'
+import beta from './modules/finally-try'
 beta.init()
 
 const Masonry = require('masonry-layout')
